# packages/sdk-ts/openapi.json
{
  "components": {
    "schemas": {
...
      "ToolCallRequest": {
        "properties": { "arguments": {} },
        "required": ["arguments"],
        "type": "object"
      },
...
      "ToolResult": {
        "properties": {
          "content": { "type": "string" },
          "is_error": { "type": "boolean" },
          "metadata": {}
        },
        "required": ["content", "is_error"],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AAgent HTTP/WS API 服务",
    "license": { "identifier": "MIT", "name": "MIT" },
    "title": "aa-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tools/{name}": {
      "post": {
        "operationId": "call_tool",
        "parameters": [
          {
            "description": "Tool name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ToolCallRequest" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ToolResult" } }
            },
            "description": "Tool execution result"
          },
          "404": {
            "content": { "text/plain": { "schema": { "type": "string" } } },
            "description": "Tool not found"
          },
          "500": {
            "content": { "text/plain": { "schema": { "type": "string" } } },
            "description": "Tool execution failed"
          }
        },
        "tags": ["AAgent"]
      }
    }
  },
...
}